refactor(lib): move color-printing functionality to CLI
