+++ crates/daktilo_lib/src/config.rs
@@ -1,10 +1,8 @@
 use crate::error::{Error, Result};
-use colored::*;
 use rdev::Key;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 use std::env;
-use std::fmt;
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::str;
@@ -114,42 +112,6 @@ pub struct SoundPreset {
     pub disabled_keys: Option<Vec<Key>>,
     /// Configure sound variations.
     pub variation: Option<SoundVariation>,
-}
-
-impl fmt::Display for SoundPreset {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        writeln!(f, "[{}]", self.name.white().bold())?;
-        let mut table = format!(
-            " {:<20}  {:<20}  {:<20}\n",
-            "Event".bold(),
-            "Keys".bold(),
-            "File".bold()
-        );
-        table.push_str(&format!(
-            " {:<20}  {:<20}  {:<20}\n",
-            "-----", "----", "----"
-        ));
-        for key_config in &self.key_config {
-            let event_str = match key_config.event {
-                KeyEvent::KeyPress => "Key Press",
-                KeyEvent::KeyRelease => "Key Release",
-            };
-            let keys_str = key_config.keys.as_str();
-            let file_str = &key_config
-                .files
-                .iter()
-                .map(|v| v.path.clone())
-                .collect::<Vec<String>>()
-                .join(",");
-            table.push_str(&format!(
-                " {:<20}  {:<20}  {:<20}\n",
-                event_str,
-                keys_str,
-                file_str.italic()
-            ));
-        }
-        write!(f, "{}", table)
-    }
 }
 
 /// Key configuration.
